pipeline/fetch.py
"""
fetch.py: Fetch raw economic events from whitelisted sources
"""

import yaml
import logging
from datetime import datetime
from pathlib import Path
from typing import Dict, List, Any

logger = logging.getLogger(__name__)


def load_sources() -> Dict[str, Any]:
    """Load allowed sources from sources.yaml"""
    sources_path = Path(__file__).parent.parent / "data" / "sources.yaml"
    with open(sources_path) as f:
        return yaml.safe_load(f)


def fetch_events(date: str) -> List[Dict[str, Any]]:
    """
    Fetch economic events for the given date from whitelisted sources.
    
    Args:
        date: YYYY-MM-DD format
    
    Returns:
        List of event dicts with source, time, description, etc.
    
    TODO:
    - Implement API integrations (Fed calendar, BLS, etc.)
    - Cache results to avoid duplicate API calls
    - Handle rate limiting & retries
    - Validate against sources.yaml
    """
    logger.info(f"Fetching events for {date}")
    
    sources = load_sources()
    events = []
    
    # Placeholder: in production, call actual APIs
    # Example integration points:
    # - Federal Reserve calendar API
    # - FRED API (economic data)
    # - NewsAPI / AlphaVantage (headline data)
    # - Manual curation (YAML input)
    
    return events


def save_raw_events(date: str, events: List[Dict[str, Any]]) -> Path:
    """Save fetched events to data/events/YYYY-MM-DD.yaml"""
    events_dir = Path(__file__).parent.parent / "data" / "events"
    output_path = events_dir / f"{date}.yaml"
    
    with open(output_path, 'w') as f:
        yaml.dump({
            'date': date,
            'timestamp': datetime.now().isoformat(),
            'event_count': len(events),
            'events': events
        }, f, default_flow_style=False)
    
    logger.info(f"Saved {len(events)} raw events to {output_path}")
    return output_path


if __name__ == "__main__":
    # Manual test
    import sys
    date = sys.argv[1] if len(sys.argv) > 1 else datetime.now().strftime("%Y-%m-%d")
    events = fetch_events(date)
    save_raw_events(date, events)
