requirements.txt
# Python dependencies for econ-llm pipeline

# YAML parsing
pyyaml>=6.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=0.19.0

# Market data & calendars
# Optional: for Yahoo Finance integration
# yfinance>=0.2.0

# Optional: Azure OpenAI support
# azure-openai>=1.0.0

# Testing & logging
pytest>=7.0
python-dateutil>=2.8.0

# HTTP requests (for API calls)
requests>=2.28.0
